% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack Solver}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with 'v' (values) and 'w' (weights) variables.}

\item{W}{The knapsack size (maximum weight).}
}
\value{
A list containing the maximum knapsack value and indices of selected items.
}
\description{
This function solves the knapsack problem using a greedy approach by selecting items based on
their value-to-weight ratio.
}
\examples{
\dontrun{
RNGversion(min(as.character(getRversion()), "3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(
  w = sample(1:4000, size = n, replace = TRUE),
  v = runif(n = n, 0, 10000)
)
result <- greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
cat("Maximum Value:", result$value, "\n")
cat("Selected Elements:", result$elements, "\n")
}

}
