% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming Knapsack Solver}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame with 'v' (values) and 'w' (weights) variables.}

\item{W}{The knapsack size (maximum weight).}
}
\value{
A list containing the maximum knapsack value and a list of selected items with their values and weights.
}
\description{
This function solves the knapsack problem using dynamic programming.
}
\examples{
\dontrun{
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(
  w = sample(1:4000, size = n, replace = TRUE),
  v = runif(n = n, 0, 10000)
)
result <- knapsack_dynamic(x = knapsack_objects[1:800,], W = 3500)
cat("Maximum Value:", result$max_value, "\n")
cat("Selected Items:", result$selected_items, "\n")
}

}
